pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nb"]

[tool.hatch.build.targets.sdist.force-include]
"docs/build/html" = "docs/build/html"

[tool.hatch.build.targets.wheel.force-include]
"docs/build/html" = "nb/_docs/"

[project]
name = "nb-cli"
version = "0.2.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8", # fuzzy date parsing
    "python-frontmatter>=1.0", # yaml frontmatter
    "rich>=13.0", # pretty printing in CLI
    "localvectordb",
    "psutil>=7.1.3",
]
[project.scripts]
nb = "nb.cli:main"
nbt = "nb.cli:main_todo"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
    "types-pyyaml>=6.0.12.20250915",
    "types-python-dateutil>=2.9.0.20251115",
    "isort>=7.0.0",
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
]
recorder = [
    "sounddevice>=0.4.6",
    "soundfile>=0.12.0",
    "numpy>=1.24.0",
    "deepgram-sdk>=3.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["nb"]
omit = ["nb/__pycache__/*", "tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.uv.sources]
localvectordb = { path = "wheels/localvectordb-1.0.0-py3-none-any.whl" }


[tool.bumpversion]
current_version = "0.2.2"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
commit = true
tag = true
tag_name = "v{new_version}"
message = "Bump version: {current_version} → {new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "nb/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'


[[tool.bumpversion.files]]
filename = "docs/source/conf.py"
search = 'release = "{current_version}"'
replace = 'release = "{new_version}"'


[tool.mypy]
python_version = "3.13"
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "frontmatter",
    "sounddevice",
    "soundfile",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py313"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (common in Click)
    "RUF005", # list concatenation (intentional for SQL params and readability)
]

[tool.ruff.lint.isort]
known-first-party = ["nb"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "F841",   # unused variable (common in test fixtures)
    "E741",   # ambiguous variable name
    "B011",   # assert False
    "PTH123",
    "RUF059"
]
